requirements.txt
fastapi
openai
passlib
psycopg2-binary
pydantic[email]
python-dotenv
requests
sqlalchemy
sse-starlette
uvicorn
